# packages/server/openapi.yaml
openapi: 3.1.0
info:
  title: agentic-kanban REST API
  version: 0.1.9
  description: "Auto-generated from Hono route source via scripts/generate-openapi.ts. Schemas are inferred statically (best-effort) — request bodies come from the parseJsonBody<T> type argument; response bodies are untyped. Coverage is audited: see x-coverage for what is scanned and what is deliberately not described here."
servers:
  - url: /
    description: "Same-origin (default dev: http://localhost:3001)"
...
paths:
...
  /api/projects:
    get:
      summary: GET /api/projects  (?includeArchived=true to include archived projects)
      tags:
        - projects
      operationId: get_api_projects
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
    post:
      summary: POST /api/projects
      tags:
        - projects
      operationId: post_api_projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repoPath:
                  type: string
                cloneUrl:
                  type: string
                name:
                  type: string
                description:
                  type: string
                color:
                  type: string
                gitignoreTemplate:
                  type: string
                generateReadme:
                  type: boolean
                exportSkillsOnRegistration:
                  type: boolean
                progressId:
                  type: string
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema: {}
        "400":
          description: Error (thrown BAD_REQUEST, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
        "409":
          description: Error (thrown CONFLICT, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
...
  /api/projects/{id}/board-risk-digest:
    get:
      summary: GET /api/projects/:id/board-risk-digest
      tags:
        - project-analytics
      operationId: get_api_projects_id_board-risk-digest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "404":
          description: Error (thrown NOT_FOUND, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
...
  /api/projects/{id}/check-overlap:
    post:
      summary: POST /api/projects/:id/check-overlap — check for file overlap between issues using cached predictions
      tags:
        - project-analytics
      operationId: post_api_projects_id_check-overlap
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "400":
          description: Error (thrown HTTPException, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
...
  /api/projects/{id}/dashboard/throughput-by-provider:
    get:
      summary: Returns count + median lead time per provider.
      tags:
        - project-analytics
      operationId: get_api_projects_id_dashboard_throughput-by-provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
  /api/projects/{id}/dependency-waves:
    get:
      summary: GET /api/projects/:id/dependency-waves
      tags:
        - project-analytics
      operationId: get_api_projects_id_dependency-waves
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
  /api/projects/{id}/dependency-waves/start-next:
    post:
      summary: POST /api/projects/:id/dependency-waves/start-next
      tags:
        - project-analytics
      operationId: post_api_projects_id_dependency-waves_start-next
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "400":
          description: Error (thrown BAD_REQUEST, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
        "404":
          description: Error (thrown NOT_FOUND, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
  /api/projects/{id}/file-contention:
    get:
      summary: GET /api/projects/:id/file-contention — live file contention heatmap for active/reviewing workspaces
      tags:
        - project-analytics
      operationId: get_api_projects_id_file-contention
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "404":
          description: Error (thrown NOT_FOUND, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
...
  /api/projects/{id}/monitor-cycles:
    get:
      summary: GET /api/projects/:id/monitor-cycles — aggregated cycle summaries
      tags:
        - project-analytics
      operationId: get_api_projects_id_monitor-cycles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
...
  /api/projects/{id}/sprint-capacity:
    get:
      summary: GET /api/projects/:id/sprint-capacity
      tags:
        - project-analytics
      operationId: get_api_projects_id_sprint-capacity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
...
  /api/projects/{id}/workspace-launch-failures:
    get:
      summary: GET /api/projects/:id/workspace-launch-failures
      tags:
        - project-analytics
      operationId: get_api_projects_id_workspace-launch-failures
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "404":
          description: Error (thrown NOT_FOUND, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
...
  /api/projects/{id}/workspace-risk:
    get:
      summary: GET /api/projects/:id/workspace-risk — risk heatmap for active/review workspaces
      tags:
        - project-analytics
      operationId: get_api_projects_id_workspace-risk
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "404":
          description: Error (thrown NOT_FOUND, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
...